requirements.txt
alembic==1.14.0
chromadb==0.5.23s
fastapi==0.115.6
loguru==0.7.3
openai==1.58.1
pdfplumber==0.11.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.34.0
groq